Add script for generating bloom filter on backend (#2)
* Add filter generator
* Tidy use of const, imports and for loop types
* Tidy up and use more appropriate size_t over unsigned int
* Match extensions error rate
* Add license agreement
* Update libraries to work on linux
* Update Catch lib to use our submodule
* Tidy up terminology
* Add more logging
* Switch to gcc

+++ src/BloomFilter.hpp
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2018 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include <iostream>
 #include <string>
 #include <vector>
@@ -16,11 +32,11 @@ typedef basic_ostream<BlockType> BinaryOutputStream;
 class BloomFilter {
 
 public:
-    BloomFilter(unsigned int maxItems, double targetProbability);
+    BloomFilter(size_t maxItems, double targetProbability);
 
-    BloomFilter(string importFilePath, unsigned int maxItems);
+    BloomFilter(string importFilePath, size_t maxItems);
 
-    BloomFilter(BinaryInputStream &in, unsigned int maxItems);
+    BloomFilter(BinaryInputStream &in, size_t maxItems);
 
     void add(string element);
 
@@ -31,6 +47,6 @@ public:
     void writeToStream(BinaryOutputStream &out);
 
 private:
-    unsigned int hashRounds;
+    size_t hashRounds;
     vector<bool> bloomVector;
 };
